refactor: ディレクトリ構成を整理
