--- barcodeokapi/lib/ui/count/count_page.dart
@@ -1,26 +1,0 @@
-import 'package:barcodeokapi/ui/count/counter.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-class Home extends ConsumerWidget {
-  @override
-  Widget build(context, ref) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('Counter example')),
-      body: Center(
-        // Consumer is a widget that allows you reading providers.
-        child: Consumer(
-          builder: (context, ref, _) {
-            final count = ref.watch(counterProvider);
-            return Text('$count');
-          },
-        ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        // The read method is a utility to read a provider without listening to it
-        onPressed: () => ref.read(counterProvider.notifier).increment(),
-        child: const Icon(Icons.add),
-      ),
-    );
-  }
-}
