+++ barcodeokapi/lib/ui_layer/resize_page.dart
